:gem: Full cleanup

+++ src/tk/wurst_client/features/commands/DamageCmd.java
@@ -1,6 +1,6 @@
 /*
  * Copyright © 2014 - 2017 | Wurst-Imperium | All rights reserved.
- * 
+ *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -41,14 +41,12 @@ public class DamageCmd extends Cmd
 		NetHandlerPlayClient sendQueue = mc.player.connection;
 		
 		// apply damage
-		for(int i = 0; (double)i < 80 + 20 * (dmg - 1D); ++i)
+		for(int i = 0; i < 80 + 20 * (dmg - 1D); ++i)
 		{
-			sendQueue.sendPacket(new Position(posX,
-				posY + 0.049D, posZ, false));
-			sendQueue.sendPacket(new Position(posX, posY,
-				posZ, false));
+			sendQueue
+				.sendPacket(new Position(posX, posY + 0.049D, posZ, false));
+			sendQueue.sendPacket(new Position(posX, posY, posZ, false));
 		}
-		sendQueue.sendPacket(new Position(posX, posY, posZ,
-			true));
+		sendQueue.sendPacket(new Position(posX, posY, posZ, true));
 	}
 }
